+++ lib/meshchat/network.rb
@@ -1,0 +1,8 @@
+module MeshChat
+  module Network
+    extend ActiveSupport::Autoload
+    eager_autoload do
+      autoload :Dispatcher
+    end
+  end
+end
