+++ lib/meshchat/ui.rb
@@ -1,0 +1,9 @@
+module MeshChat
+  module Ui
+    extend ActiveSupport::Autoload
+    autoload :CLI
+    autoload :Command
+    autoload :Display
+    autoload :Notifier
+  end
+end
