+++ lib/meshchat/ui/notifier.rb
@@ -1,0 +1,8 @@
+module MeshChat
+  module Ui
+    module Notifier
+      extend ActiveSupport::Autoload
+      autoload :Base
+    end
+  end
+end
